[REFACTOR][E] Rename package "saros.project"
Renames the package "saros.project" to "saros.resource_change_handlers"
to better represent the contained classed.

Fixes the container package-info (previously incorrectly named
"project-info").
