+++ eclipse/META-INF/MANIFEST.MF
@@ -31,7 +31,7 @@ Export-Package: saros;
    saros.net.internal,
    org.jivesoftware.smackx.muc,
    org.jivesoftware.smack,
-   saros.project,
+   saros.resource_change_handlers,
    org.eclipse.core.resources,
    org.osgi.framework,
    org.jivesoftware.smackx,
@@ -42,7 +42,7 @@ Export-Package: saros;
    org.eclipse.core.runtime,
    saros.activities,
    org.eclipse.ui,
-   saros.project,
+   saros.resource_change_handlers,
    saros.editor.internal,
    org.xmlpull.v1,
    saros.invitation,
@@ -57,7 +57,7 @@ Export-Package: saros;
  saros.filesystem,
  saros.monitoring,
  saros.preferences,
- saros.project;
+ saros.resource_change_handlers;
   uses:="saros.activities,
    saros,
    org.jivesoftware.smack,
@@ -77,7 +77,7 @@ Export-Package: saros;
    saros.invitation,
    org.eclipse.jface.preference,
    org.jivesoftware.smack,
-   saros.project,
+   saros.resource_change_handlers,
    org.eclipse.swt.graphics,
    org.eclipse.jface.util,
    org.eclipse.swt.widgets,
@@ -88,7 +88,7 @@ Export-Package: saros;
    saros,
    org.jivesoftware.smack,
    org.eclipse.ui,
-   saros.project,
+   saros.resource_change_handlers,
    org.eclipse.ui.actions,
    org.eclipse.jface.viewers,
    saros.invitation,
@@ -96,7 +96,7 @@ Export-Package: saros;
  saros.ui.command_handlers,
  saros.ui.decorators;
   uses:="org.eclipse.core.runtime,
-   saros.project,
+   saros.resource_change_handlers,
    org.eclipse.jface.viewers,
    saros.editor,
    saros.invitation",
