--- eclipse/src/saros/project/project-info.java
@@ -1,10 +1,0 @@
-/**
- * The classes in the project package are responsible for managing the shared session and shared
- * projects.
- *
- * <p>The session is represented by {@link ISarosSession} and the {@link ISarosSessionManager}
- * controls its lifecycle. A {@link SharedProject} belongs to a session. Additionally, shared
- * resources are monitored within this package and the framework for processing activities is laid
- * out here.
- */
-package saros.project;
