+++ eclipse/src/saros/resource_change_handlers/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains all handlers for local and remote resource changes. Changes through the
+ * editor are explicitly excluded as they are handled by the {@link saros.editor.EditorManager}.
+ */
+package saros.resource_change_handlers;
