+++ eclipse/test/junit/saros/SarosEclipseTestSuite.java
@@ -7,7 +7,7 @@ import org.junit.runners.Suite;
 @Suite.SuiteClasses({
   saros.concurrent.undo.TestSuite.class,
   saros.editor.internal.TestSuite.class,
-  saros.project.TestSuite.class,
+  saros.resource_change_handlers.TestSuite.class,
   saros.session.internal.TestSuite.class,
   saros.ui.model.roster.TestSuite.class,
   saros.util.UtilTestSuite.class,
