+++ eclipse/test/junit/saros/resource_change_handlers/TestSuite.java
@@ -1,4 +1,4 @@
-package saros.project;
+package saros.resource_change_handlers;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
